Create ServiceItemDetails component
