+++ src/components/ServiceItemTableHasImg/index.module.scss
@@ -22,36 +22,10 @@
   }
 }
 
-.detailLine {
-  @include u-font-size('body', '3xs');
-}
-
-// abstract called for thin semi bold which isn't a variant
-.detailType {
-  @include u-font-weight(600);
-}
-
-.detailImage {
-  @include u-display(flex);
-}
-
-.siThumbnail {
-  width: 100px;
-  height: 100px;
-}
-
 .statusAction {
   @include u-display(flex);
 
   button {
     @include u-margin-top(0);
   }
-}
-
-.customerContact {
-  @include u-margin-top(2);
-}
-
-.detailImage p {
-  @include u-margin-top(0);
 }
